seed/requirements.txt
psycopg[binary]==3.2.2
Faker==25.8.0
python-dotenv==1.0.1